P0 optimization: dead code removal, memory reduction, binary size
Dead code removed (16 items):
- Delete entire Accessibility class (accessibility.h/.cpp) — 0 callers
- Remove AnimatedProperty enum (animation.h) — 0 references
- Remove Widget::minimumSize() (widget.h/cpp) — 0 callers
- Remove Button::animDuration_ field (button.h) — never read/written
- Remove Window::imeCursorScreenPos() (window.h/cpp) — 0 callers
- Remove TextBox::cursorBytePos/VisualLine/VisualCol/posFromVisualPos
  (textbox.h) — methods declared but never defined or called
- Remove FontAtlas::flush()/atlasTexId() — 0 callers
- Remove Renderer2D::queuedCommands() (gpu_renderer.h) — 0 callers
- Remove GpuCanvas::defaultFontData_[256] (gpu_canvas.h) — 0 reads
- Remove 14 unused Color constants (color.h/cpp) — only tests referenced
- Remove #include <animation.h> from button.h — unused
- Remove #include <cmath> from progressbar.cpp — unused
- Remove DialogButton::Close enum value — 0 code paths use it

Memory reduction (2 items):
- Widget: 5 bool fields → 1 uint8_t bitfield flags_ (saves ~12 bytes/widget)
- Style: 8 int padding/margin → int16_t (saves ~16 bytes/Style)

Binary size reduction (2 items):
- Color::toARGB() kept (useful API, 1 line), Color constants reduced 14→4

Widget infrastructure:
- Widget::minimumSize() removed — unused virtual
- Widget::sizeHintDirty_/needsLayout_ migrated to flags_ bitfield
- Widget::setCachedSizeHint/sizeHint const-correctness preserved via mutable

Test updates:
- test_color.cpp: removed tests for deleted Color constants
- test_integration.cpp: use Button instead of base Widget for focus tests
  (base Widget::canAcceptFocus() now correctly returns false)

17/17 tests pass. 37 files, +446/-448 (net -2 lines)

+++ include/ltgui/style.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "color.h"
 #include "font.h"
+#include <cstdint>
 #include <algorithm>
 
 namespace ltgui {
@@ -12,14 +13,14 @@ struct Style {
     int borderWidth = 0;
     int borderRadius = 0;
     Font font;
-    int paddingLeft = 0;
-    int paddingTop = 0;
-    int paddingRight = 0;
-    int paddingBottom = 0;
-    int marginLeft = 0;
-    int marginTop = 0;
-    int marginRight = 0;
-    int marginBottom = 0;
+    int16_t paddingLeft = 0;
+    int16_t paddingTop = 0;
+    int16_t paddingRight = 0;
+    int16_t paddingBottom = 0;
+    int16_t marginLeft = 0;
+    int16_t marginTop = 0;
+    int16_t marginRight = 0;
+    int16_t marginBottom = 0;
 
     static Style defaultStyle();
